Revert "clk: clk-mux: Fold clk_register_mux()"
clk_register_mux_table() is used by a few linux clk drivers. It's very
useful for the muxes that only few values are valid. It should also be
provided by U-Boot CLK CCF framework to be consistant with linux kernel.

This reverts commit 7cc520fe297dba17f681377ad201f258f927c662.

Revert this commit to reintroduce this function back. It will be used by
the upcoming moson8b clk driver.